+++ app/src/main/java/maitre/app/layout/MainActivity.kt
@@ -4,6 +4,7 @@ import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import maitre.app.R
+import maitre.app.Reserva
 import maitre.app.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
