+++ app/src/main/java/maitre/app/utils/Endpoints.kt
@@ -6,12 +6,11 @@ import retrofit2.http.Body
 import retrofit2.http.GET
 import retrofit2.http.POST
 import retrofit2.http.Path
-import java.util.Optional
 
 interface Endpoints {
 
     @GET("/usuarios/{email}/{senha}")
-    fun getLogin(@Path("email") email: String, @Path("senha") senha: String) : Call<Optional<Usuario>>
+    fun getLogin(@Path("email") email: String, @Path("senha") senha: String) : Call<Usuario>
 
     @POST("/usuarios")
     fun cadastrar(@Body usuario: Usuario) : Call<Usuario>
