+++ app/src/main/res/layout/activity_home.xml
@@ -4,7 +4,8 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Home"
+    tools:context=".layout.Home"
+    android:id="@+id/layout_home"
     android:background="@mipmap/background_home">
