+++ app/src/main/res/layout/fragment_confirmacao_reserva.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ConfirmacaoReserva">
+    tools:context=".layout.ConfirmacaoReserva">
 
     <RelativeLayout
         android:layout_width="match_parent"
