+++ app/src/main/res/layout/fragment_fila.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Fila">
+    tools:context=".layout.Fila">
 
     <RelativeLayout
         android:layout_width="match_parent"
