+++ app/src/main/res/layout/fragment_lgpd.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Lgpd">
+    tools:context=".layout.Lgpd">
 
 <!--   Principal-->
     <RelativeLayout
