+++ app/src/main/res/layout/fragment_telefone_estabelecimento.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".TelefoneEstabelecimento">
+    tools:context=".layout.TelefoneEstabelecimento">
 
     <RelativeLayout
         android:layout_width="match_parent"
